+++ tests/nn/data_parallel/test_fsdp.py
@@ -27,6 +27,7 @@ from fairscale.utils.testing import (
     dist_init,
     get_cycles_per_ms,
     objects_are_equal,
+    skip_a_test_if_in_CI,
     spawn_for_all_world_sizes,
 )
 
@@ -480,11 +481,12 @@ class TestReduceScatterProcessGroup(DistributedTest):
             return FullyShardedDataParallel(model, group, **config)
 
 
-@pytest.mark.skip(reason="Disable flaky test that is not reproducible locally.")
+@pytest.mark.skip(reason="Recently flaky and not reproducible locally.")
 class TestSerialization(DistributedTest):
     @parameterized.expand([[False, False], [True, False], [True, True], [False, True]], name_func=rename_test)
     def test_pickle(self, mixed_precision, cpu_offload):
         """Ensure that wrapped modules can be pickled/unpickled."""
+        skip_a_test_if_in_CI()
         config = {"mixed_precision": mixed_precision, "cpu_offload": cpu_offload}
         test_fn = functools.partial(self._test_pickle, config=config)
         spawn_and_init(test_fn, world_sizes=[2])
@@ -492,6 +494,7 @@ class TestSerialization(DistributedTest):
     @parameterized.expand([[False, False], [True, False], [True, True], [False, True]], name_func=rename_test)
     def test_multiprocessing(self, mixed_precision, cpu_offload):
         """Ensure that wrapped modules can be sent via multiprocessing."""
+        skip_a_test_if_in_CI()
         config = {"mixed_precision": mixed_precision, "cpu_offload": cpu_offload}
         test_fn = functools.partial(self._test_multiprocessing, config=config)
         spawn_and_init(test_fn, world_sizes=[2])
@@ -772,7 +775,7 @@ class DummyDDP(nn.Module):
 
 
 class MixtureOfExperts(NestedWrappedModule):
-    def __init__(self, group, wrapper_config, checkpoint_act=False, delay_before_free_ms=0):
+    def __init__(self, group, wrapper_config, checkpoint_act=False, delay_before_free_ms=0, expert_group=None):
         super().__init__(group, wrapper_config)
         self.group = group
         self.delay_before_free_ms = delay_before_free_ms
@@ -798,9 +801,12 @@ class MixtureOfExperts(NestedWrappedModule):
             shared = checkpoint_wrapper(shared)
 
         if wrapper_config is not None:
-            # we create a process group of size 1 for the expert params
-            expert_group = torch.distributed.new_group([group.rank()])  # world size 1 means no shard
-            expert = FullyShardedDataParallel(expert, expert_group, **wrapper_config)
+            # we create a process group of size >= 1 for the expert params
+            # we also need to pass that group as the reduce_scatter group.
+            expert_group = expert_group or torch.distributed.new_group([group.rank()])
+            expert = FullyShardedDataParallel(
+                expert, process_group=expert_group, process_group_reduce_scatter=expert_group, **wrapper_config
+            )
 
             shared = FullyShardedDataParallel(shared, group, **wrapper_config)