+++ tests/nn/data_parallel/test_fsdp_optimizer_utils.py
@@ -4,6 +4,7 @@
 # LICENSE file in the root directory of this source tree.
 import functools
 from time import time
+import unittest
 
 from parameterized import parameterized
 import torch
@@ -12,7 +13,7 @@ from torch.optim import SGD, Adadelta, Adam  # type: ignore
 from fairscale.nn import FullyShardedDataParallel
 from fairscale.nn.data_parallel.fsdp_optim_utils import is_singleton_tensor
 from fairscale.utils.params import recursive_copy_to_device
-from fairscale.utils.testing import objects_are_equal
+from fairscale.utils.testing import dist_init, objects_are_equal, spawn_for_all_world_sizes
 
 from .test_fsdp import (
     DistributedTest,
@@ -24,15 +25,68 @@ from .test_fsdp import (
 )
 
 
-def first_tensor_numel(dct):
+def all_tensors_numel_except_for_step(dct):
+    """Compute the sum of numel from all tensors from a dict, except when the key is `step`."""
+    ret = 0
     for k, v in dct.items():
-        if torch.is_tensor(v):
-            return v.numel()
-    return 0
+        if k != "step" and torch.is_tensor(v):
+            ret += v.numel()
+    return ret
 
 
 def assert_equal(a, b):
     assert a == b, f"{a} != {b}"
+
+
+def spawn_and_init_multiple_groups(fn, args=None, **spawn_kwargs):
+    if args is None:
+        args = ()
+
+    run_fn = functools.partial(init_and_run, fn, args)
+    spawn_for_all_world_sizes(run_fn, **spawn_kwargs)
+
+
+def _find_my_group_index(grouped_ranks):
+    """Return the index corresponding to the MoE group of the current process."""
+    my_rank = torch.distributed.get_rank()
+    for i, group in enumerate(grouped_ranks):
+        if my_rank in group:
+            return i
+    raise RuntimeError(f"Unable to find process rank {my_rank} in the set of grouped ranks {grouped_ranks}.")
+
+
+def get_moe_group(moe_expert_count=2):
+    """Return a process group for initializing a MoE layer."""
+    if torch.distributed.is_initialized():
+        world_size = torch.distributed.get_world_size()
+
+        # If you have more experts than the world size.
+        if world_size <= moe_expert_count:
+            assert moe_expert_count % world_size == 0
+            moe_groups = [[i] for i in range(world_size)]
+
+        # If you have a larger world size than experts.
+        else:
+            assert world_size % moe_expert_count == 0
+            ranks_per_group = world_size // moe_expert_count
+            moe_groups = [[i + j * moe_expert_count for j in range(ranks_per_group)] for i in range(moe_expert_count)]
+
+        moe_pgs = [torch.distributed.new_group(g) for g in moe_groups]
+
+        # Find the index in the set of moe_groups which contains the current rank.
+        my_group_idx = _find_my_group_index(moe_groups)
+        return moe_pgs[my_group_idx]
+    else:
+        return torch.distributed.new_group([torch.distributed.get_rank()])
+
+
+def init_and_run(fn, args, rank, world_size, filename, filename_rpc):
+    """Initialize and run the unit test for testing replicated MoE groups."""
+    dist_init(rank, world_size, filename, filename_rpc)
+    torch.cuda.set_device(rank)
+    group = torch.distributed.new_group()
+    # Specify the moe_group used to initialize the MoE layers with.
+    fn(rank, group, *args, expert_group=get_moe_group())
 
 
 class TestOptimizerUtils(DistributedTest):
@@ -49,17 +103,33 @@ class TestOptimizerUtils(DistributedTest):
 
         spawn_and_init(test_fn, world_sizes=[min(torch.cuda.device_count(), 4)])
 
+    @parameterized.expand(
+        [[SGD, False], [Adam, False]],
+        name_func=rename_test,
+    )
+    def test_consolidate_optimizer_diff_world_size(self, optim_fn, transformer):
+        if torch.cuda.device_count() < 4:
+            raise unittest.SkipTest("This test requires at least 4 GPUs.")
+        config = {"mixed_precision": True, "flatten_parameters": True}
+        config["compute_dtype"] = torch.float32
+        test_fn = functools.partial(self._test_consolidated_optimizer, config, optim_fn=Adam, transformer=transformer)
+
+        spawn_and_init_multiple_groups(test_fn, world_sizes=[min(torch.cuda.device_count(), 4)])
+
     @classmethod
-    def _test_consolidated_optimizer(self, config, rank, group, optim_fn=torch.optim.SGD, transformer=False):
+    def _test_consolidated_optimizer(
+        self, config, rank, group, optim_fn=torch.optim.SGD, transformer=False, expert_group=None
+    ):
         """FSDP.gather_full_optim_state_dict() should return something very similar to optimizer.state_dict()"""
         # Establish reference behavior.
-
         if transformer:
             unwrapped_model = TransformerWithSharedParams(group, wrapper_config=config).cuda()
             fsdp = self.get_wrapped_model(group, config=config).cuda()
         else:
-            unwrapped_model = MixtureOfExperts(group, wrapper_config=None).cuda()
-            fsdp = FullyShardedDataParallel(MixtureOfExperts(group, wrapper_config=config)).cuda()
+            unwrapped_model = MixtureOfExperts(group, wrapper_config=None, expert_group=expert_group).cuda()
+            fsdp = FullyShardedDataParallel(
+                MixtureOfExperts(group, wrapper_config=config, expert_group=expert_group)
+            ).cuda()
 
         try:
             fsdp_optim = optim_fn(
@@ -86,9 +156,9 @@ class TestOptimizerUtils(DistributedTest):
             optim_unwrapped.step()
         unwrapped_sd = optim_unwrapped.state_dict()
 
-        if not transformer:
+        if not transformer and not expert_group:
             no_broadcast_children = [x for x in fsdp._fsdp_instances if x.no_broadcast_optim_state]
-            assert len(no_broadcast_children) == 1
+            assert len(no_broadcast_children) == 1, f"Length of non shared params {len(no_broadcast_children)}"
             assert fsdp._fsdp_instances[-1].no_broadcast_optim_state
         torch.cuda.empty_cache()
         cuda_gb_before = torch.cuda.memory_stats(fsdp.rank)["allocated_bytes.all.current"] / 1024 ** 3
@@ -113,6 +183,18 @@ class TestOptimizerUtils(DistributedTest):
                     msg = f"got device {t.device} for {k}: {buffer_name}. expected CPU"
                     assert t.device == torch.device("cpu"), msg
 
+        if expert_group:
+            sd_state = recursive_copy_to_device(sd["state"], non_blocking=False, device="cpu")
+            orig_state = recursive_copy_to_device(unwrapped_sd["state"], non_blocking=False, device="cpu")
+
+            assert_equal(len(sd_state.keys()), len(orig_state.keys()))
+
+            assert_equal(
+                sum([all_tensors_numel_except_for_step(v) for k, v in sd_state.items()]),
+                sum([all_tensors_numel_except_for_step(v) for k, v in orig_state.items()]),
+            )
+            return
+
         unflat_state = sd["state"]
         assert "uncollected_local_ids" in sd
         shard_sd = fsdp.get_shard_from_optim_state_dict(sd)
@@ -123,8 +205,8 @@ class TestOptimizerUtils(DistributedTest):
         assert_equal(len(sd["state"]), len(unwrapped_sd["state"]))
         assert_equal(len(sd["param_groups"][0]["params"]), len(unwrapped_sd["param_groups"][0]["params"]))
         assert_equal(
-            sum([first_tensor_numel(v) for k, v in sd["state"].items()]),
-            sum([first_tensor_numel(v) for k, v in unwrapped_sd["state"].items()]),
+            sum([all_tensors_numel_except_for_step(v) for k, v in sd["state"].items()]),
+            sum([all_tensors_numel_except_for_step(v) for k, v in unwrapped_sd["state"].items()]),
         )
 
         original_shard_sd = fsdp_optim.state_dict()
@@ -133,8 +215,8 @@ class TestOptimizerUtils(DistributedTest):
         original_shard_sd = recursive_copy_to_device(original_shard_sd, non_blocking=False, device="cpu")
         # Before asserting that the dicts are equal, we check keys individually to allow nice tracebacks.
         assert_equal(
-            [first_tensor_numel(v) for k, v in shard_sd["state"].items()],
-            [first_tensor_numel(v) for k, v in original_shard_sd["state"].items()],
+            [all_tensors_numel_except_for_step(v) for k, v in shard_sd["state"].items()],
+            [all_tensors_numel_except_for_step(v) for k, v in original_shard_sd["state"].items()],
         )
         assert_equal(
             [v for k, v in shard_sd["param_groups"][0].items()],