+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   class separation margins and feature vector scaling to the standard
   output projection + cross-entropy loss. MEVO supported this with its
   memory saving techniques so that peak GPU memory is much reduced. [#916]
+- FSDP: Added skip_params_check_for_root flag to default_auto_wrap_policy which,
+  if set, wraps the root module regardless of how many unwrapped params there were
+  left after children were wrapped. [#930]
+- FSDP: Add support for saving optimizer state when using expert replicas with FSDP.
 
 ## [0.4.5] - 2022-01-14